Extract server building into startup module
